+++ packages/apollo-boost/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### vNext
 
+- Allow `cache` to be given as a configuration option to `ApolloBoost`.
+  [PR #3561](https://github.com/apollographql/apollo-client/pull/3561)
 - Allow `headers` and `credentials` to be passed in as configuration
   parameters to the `apollo-boost` `ApolloClient` constructor.
   [PR #3098](https://github.com/apollographql/apollo-client/pull/3098)